backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
motor==3.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cloudinary==1.37.0
stripe==6.1.0
python-dotenv==1.0.0
boto3==1.29.0
alembic==1.12.1
sqlalchemy==2.0.23
pytest==7.4.3
pytest-asyncio==0.21.1
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.0.0.post2
